focal-backend/requirements.txt
asgiref==3.3.4
bcrypt==3.2.0
cffi==1.14.5
click==8.0.1
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.65.2
h11==0.12.0
httptools==0.2.0
idna==3.2
install==1.3.4
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
python-dotenv==0.18.0
python-jose==3.3.0
PyYAML==5.4.1
rsa==4.7.2
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.0
unicorn==1.0.3
uvicorn==0.14.0
uvloop==0.15.2
watchgod==0.7
websockets==9.1
